Fix template type parameter error

+++ KnowledgeBase.hpp
@@ -113,7 +113,7 @@ void KnowledgeBase<TRow, TCol>::ComputeLinkStrengths()
 {
 	std::unique_ptr<IKnowledgeLinks<TRow, TCol, float>> link_strengths(new SparseHashLinksMatrix<TRow, TCol, float>(cooccurrence_counts_->GetNumRows(), cooccurrence_counts_->GetNumCols()));
 
-	for (const std::pair<std::pair<TRow, TCol>, float>& e: cooccurrence_counts_->GetNzElements()) {
+	for (const std::pair<std::pair<TRow, TCol>, uint32_t>& e: cooccurrence_counts_->GetNzElements()) {
 		TRow row = e.first.first;
 		TCol col = e.first.second;
 		link_strengths->SetElement(row, col, ComputeLinkStrength(float(e.second) / (float) target_symbol_sums_[row]));
